requirements.txt
 Django>=5.2.8,<6
djangorestframework
django-cors-headers
gunicorn
numpy
Pillow
python-dotenv
psycopg2-binary
tensorflow-cpu==2.26.0
